+++ common/model/model.go
@@ -80,8 +80,8 @@ func (dt VPAType) String() string {
 
 type Query struct {
 	Predicate string `json:"predicate" required:"true"`
-	Offset    int    `json:"offset"`
-	Limit     int    `json:"limit"`
+	Offset    int64    `json:"offset"`
+	Limit     int64    `json:"limit"`
 }
 
 func initValidator() *validator.Validate {
